+++ util/updateXML/update1.8.0.6.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2025, RTE (http://www.rte-france.com)
+# See AUTHORS.txt
+# All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, you can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of Dynawo, a hybrid C++/Modelica open source suite
+# of simulation tools for power systems.
+
+from content.Ticket import ticket
+
+# GeneratorFictitious -> GeneratorAlphaBeta
+@ticket(3897)
+def update(jobs):
+    modelica_models = jobs.dyds.get_modelica_models(lambda _: True)
+    for modelica_model in modelica_models:
+        unit_dynamic_models = modelica_model.get_unit_dynamic_models(
+            lambda unit_dynamic_model: unit_dynamic_model.get_name() == "Dynawo.Electrical.Machines.Simplified.GeneratorFictitious"
+        )
+        for unit_dynamic_model in unit_dynamic_models:
+            unit_dynamic_model.set_name("Dynawo.Electrical.Machines.Simplified.GeneratorAlphaBeta")
+
+    for genxml in jobs.dyds.get_bbms(lambda bbm: bbm.get_lib_name() == "GeneratorFictitious"):
+        genxml.set_lib_name("GeneratorAlphaBeta")
